refactor(routes): add role and user status checks for routes of listingRouter

+++ backend/src/routes/listing.router.ts
@@ -1,9 +1,12 @@
 import { Router } from "express";
 
 import { listingController } from "../controllers/listing.controller";
+import { PlatformPermissionEnum } from "../enums/platform-permission.enum";
 import { TokenTypeEnum } from "../enums/token-type.enum";
 import { authMiddleware } from "../middlewares/auth.middleware";
 import { commonMiddleware } from "../middlewares/common.middleware";
+import { roleMiddleware } from "../middlewares/role.middleware";
+import { userMiddleware } from "../middlewares/user.middleware";
 
 const router = Router();
 
@@ -18,6 +21,9 @@ router.get(
 router.post(
     "/",
     authMiddleware.checkToken(TokenTypeEnum.ACCESS),
+    userMiddleware.isActiveUser,
+    userMiddleware.isVerifiedUser,
+    roleMiddleware.checkPermission(PlatformPermissionEnum.LISTING_CREATE),
     listingController.create,
 );
 
@@ -25,6 +31,9 @@ router.patch(
     "/:listingId",
     commonMiddleware.isValidId("listingId"),
     authMiddleware.checkToken(TokenTypeEnum.ACCESS),
+    userMiddleware.isActiveUser,
+    userMiddleware.isVerifiedUser,
+    roleMiddleware.checkPermission(PlatformPermissionEnum.LISTING_EDIT),
     listingController.updateById,
 );
 
@@ -32,6 +41,9 @@ router.patch(
     "/:listingId/activate",
     commonMiddleware.isValidId("listingId"),
     authMiddleware.checkToken(TokenTypeEnum.ACCESS),
+    userMiddleware.isActiveUser,
+    userMiddleware.isVerifiedUser,
+    roleMiddleware.checkPermission(PlatformPermissionEnum.LISTING_ACTIVATE),
     listingController.activateListing,
 );
 
@@ -39,6 +51,9 @@ router.patch(
     "/:listingId/deactivate",
     commonMiddleware.isValidId("listingId"),
     authMiddleware.checkToken(TokenTypeEnum.ACCESS),
+    userMiddleware.isActiveUser,
+    userMiddleware.isVerifiedUser,
+    roleMiddleware.checkPermission(PlatformPermissionEnum.LISTING_DEACTIVATE),
     listingController.deactivateListing,
 );
 
@@ -46,13 +61,19 @@ router.delete(
     "/:listingId",
     commonMiddleware.isValidId("listingId"),
     authMiddleware.checkToken(TokenTypeEnum.ACCESS),
+    userMiddleware.isActiveUser,
+    userMiddleware.isVerifiedUser,
+    roleMiddleware.checkPermission(PlatformPermissionEnum.LISTING_DELETE),
     listingController.deleteById,
 );
 
 router.get(
     "/:listingId/statistics",
-    authMiddleware.checkToken(TokenTypeEnum.ACCESS),
     commonMiddleware.isValidId("listingId"),
+    authMiddleware.checkToken(TokenTypeEnum.ACCESS),
+    userMiddleware.isActiveUser,
+    userMiddleware.isVerifiedUser,
+    roleMiddleware.checkPermission(PlatformPermissionEnum.LISTING_GET_STATS),
     listingController.getPremiumListingStats,
 );
